refactor: keep apply buttons server-rendered; add on-page apply e2e
Address PR review:
- Extract ApplyButton + the APPLY product map into a plain (non-client)
  ui/apply-buttons module, so the buttons stay server-rendered wherever
  used. HelpMeChoose (client) and ApplyOptions (server) both import from
  it; HelpMeChoose no longer re-exports a component that would pull the
  buttons into the client bundle.
- Refresh the wizard spec header (the eligibility wizard page is retired)
  and add a Playwright test for the on-page "Help me choose" flow on
  /help-for-charities/: selecting the determination-letter answer reveals
  the pid 33 application link in place.
